+++ app/templates/unsubscribe.html
@@ -1,0 +1,15 @@
+{% load rest_framework %}
+
+<html>
+  <body>
+    <h1>Unsubscribe</h1>
+    <h2>{{ trigger.monitor.name }}</h2>
+    <form
+      action="{% url 'measurement:trigger-unsubscribe' token=token pk=pk %}"
+      method="POST"
+    >
+      {% csrf_token %}{% render_form serializer %}
+      <button type="submit">Unsubscribe</button>
+    </form>
+  </body>
+</html>
